move types and some other functions into main
